requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
openai==1.107.1
psycopg[binary]==3.2.9
pandas==2.3.2
openpyxl==3.1.5
email-validator==2.2.0
